Adds `Sentry::Config` for managing configuration; Checks `.sentry.yml` on startup
* Add `Sentry::Config` for managing configuration, and check for `.sentry.yml`.

This commit introduces `Sentry::Config` as a single object to manage the configuration of a `sentry` instance. The CLI also now checks for a `.sentry.yml` file in the project directory to load configurations automatically. These configurations can still be overwritten by specifying options the command line, though this is now entirely optional, as all options are supported from the YAML file.

An example and explanation of the `.sentry.yml` format can be found in `.sentry.example.yml`.

* Allow users to specify a config file with `-c` or `--config`

* Rename `process_name` to `display_name`. Use name from shard for default build/run commands.

* Only set `display_name` on merge if it was explicitly set on `other`.
The default getter falls back to the shard name, and because the conditional for merging `display_name` only checked that this was truthy, it would always be set to whatever `other` had as a `display_name`, or the shard name (never keeping its own value).

* Track whether config explicitly sets `build` and `run` commands.

Similar to `display_name` before, `build` and `run` were also being overwritten when merging configs because of their default values.

This commit adds more boolean properties to `Sentry::Config` to track whether the `build` and `run` commands have been explicitly set, rather than using the default value. These booleans are used in `merge!` to decide whether the config being merged in should apply its `build` and `run` commands to the receiving config.